% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_m.R
\name{ass_m}
\alias{ass_m}
\title{Create an host-symbiont association matrix}
\usage{
ass_m(hs)
}
\arguments{
\item{hs}{A two-columns matrix representing associations between host
(column 1) and symbiont (column 2) species.}
}
\value{
An association matrix, with hosts in rows and symbionts
in columns, sorted alphabetically.
}
\description{
Create an host-symbiont association matrix
}
