% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_mat.R
\name{assoc_mat}
\alias{assoc_mat}
\title{Create a host-symbiont association matrix}
\usage{
assoc_mat(hs)
}
\arguments{
\item{hs}{A two-columns matrix representing associations between host
(column 1) and symbiont (column 2) species.}
}
\value{
An association matrix, with hosts in rows and symbionts in columns,
        sorted alphabetically.
}
\description{
Create a host-symbiont association matrix
}
\examples{
data(plant_fungi)
pf_matrix <- assoc_mat(pf)

}
