% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\docType{data}
\name{birds}
\alias{birds}
\title{Passerine birds phylogeny}
\format{
An object of class \code{"phylo"} containing a list with the
details of the phylogenetic tree (i.e. edges, edges length, nodes, root edge
and tips names).
}
\source{
Balbuena, J. A., Pérez-Escobar, Ó. A., Llopis-Belenguer, C.,
Blasco-Costa, I. (2020). Random tanglegram partitions (Random TaPas):
An alexandrian approach to the cophylogenetic gordian knot. Systematic
biology, 69(6), 1212-1230. \doi{10.1093/sysbio/syaa033}.
}
\usage{
data(birds_mites)
}
\description{
Phylogeny of passerine brids extracted from one of 200 random stationary
Bayesian time-calibrated trees downloaded form "A global phylogeny of birds"
http://birdtree.org .
}
\references{
Klimov, P. B., Mironov, S. V., OConnor, B. M. (2017).
Detecting ancient codispersals and host shifts by double dating of host and
parasite phylogenies: Application in proctophyllodid feather mites associated
with passerine birds. Evolution, 71(10), 2381-2397. \doi{10.1111/evo.13309}
}
\keyword{datasets}
