% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_matrix.R
\docType{data}
\name{bm_matrix}
\alias{bm_matrix}
\title{Passerine birds associations with feather mites}
\format{
A binary matrix with 93 rows (birds) and 98 variables (feather
mites).
}
\source{
Balbuena, J. A., Pérez-Escobar, Ó. A., Llopis-Belenguer, C.,
Blasco-Costa, I. (2020). Random tanglegram partitions (Random TaPas):
An alexandrian approach to the cophylogenetic gordian knot. Systematic
biology, 69(6), 1212-1230. \doi{10.1093/sysbio/syaa033}.
}
\usage{
data(birds_mites)
}
\description{
Associations between 93 passerine birds species and their feather mite,
collected in eight regions (Costa Rica, Kazakhstan, Mexico, Panama, Peru,
Russia, Tanzania, USA) by Klimov et al. (2017).
}
\references{
Klimov, P. B., Mironov, S. V., & OConnor, B. M. (2017).
Detecting ancient codispersals and host shifts by double dating of host and
parasite phylogenies: Application in proctophyllodid feather mites associated
with passerine birds. Evolution, 71(10), 2381-2397. \doi{10.1111/evo.13309}
}
\keyword{datasets}
