% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_mask}
\alias{brain_mask}
\title{Brain Mask}
\usage{
brain_mask
}
\description{
Brain mask for all probability maps and gold standard masks.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
