% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tapas_data.R
\name{gen_tapas_data}
\alias{gen_tapas_data}
\title{Calculate Threshold Grid}
\usage{
gen_tapas_data(thresholds = seq(from = 0, to = 1, by = 0.01), pmap,
  gold_standard, mask, k = 8, subject_id = NULL, verbose = TRUE)
}
\arguments{
\item{thresholds}{A \code{vector} of thresholds to use for calculation of Sørensen's–Dice coefficient
and automatic volume. The default and grid applied in the published work is 0 to 1. Threshold
values must be between 0 and 1.}

\item{pmap}{A file path to a \code{NIFTI} probability map generated from an automatic segmentation
approach or the name of a probability map locally available}

\item{gold_standard}{A file path to the \code{NIFTI} gold standard segmentation approach corresponding
with the probability map provided or the name of the gold standard segmentation that has been
locally loaded. The gold standard segmentation is used to compare the thresholded probability map
image with using Sørensen's–Dice coefficient.}

\item{mask}{A file path to a \code{NIFTI} brain mask corresponding with the probability map provided
or the name of a mask that has been locally loaded.}

\item{k}{Minimum number of voxels for a cluster/component. Passed to \code{extransr::label_mask()}.}

\item{subject_id}{is NULL by default. Must be a \code{character} vector with the unique subject ID for
which the threshold data is being generated.}

\item{verbose}{is TRUE by default. TRUE returns messages throughout the generating data function.
FALSE will silence comment returns.}
}
\value{
A tibble with threshold, dsc, and volume columns
}
\description{
Calculates Sørensen's–Dice coefficient and automatic volume
 estimation for a grid of thresholds.
}
\examples{
\dontrun{
gen_tapas_data(thresholds = seq(from = 0, to = 1, by = 0.01),
pmap = 'probability_map_subject_1.nii.gz',
gold_standard = 'manual_segmentation_subject_1.nii.gz',
mask = 'brain_mask_subject_1.nii.gz',
k = 8,
subject_id = 'subject_1')
}
}
