% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_thresh_data.R
\name{gen_thresh_data}
\alias{gen_thresh_data}
\title{Calculate Threshold Grid}
\usage{
gen_thresh_data(thresholds = seq(from = 0, to = 1, by = 0.01), pmap,
  gold_standard, mask, k = 8)
}
\arguments{
\item{thresholds}{A vector of thresholds to use for calculation of Sørensen's–Dice coefficient
and automatic volume. The default and grid applied in the published work is 0 to 1. Threshold
values must be between 0 and 1.}

\item{pmap}{A file path to a NIFTI probability map generated from an automatic segmentation
approach or the name of a probability map that has been locally loaded}

\item{gold_standard}{A file path to the NIFTI gold standard segmentation approach corresponding
with the probability map provided or the name of the gold standard segmentation that has been
locally loaded. The gold standard segmentation is used to compare the thresholded probability map
image with using Sørensen's–Dice coefficient.}

\item{mask}{A file path to a NIFTI brain mask corresponding with the probability map provided
or the name of a mask that has been locally loaded.}

\item{k}{A file path to a NIFTI brain mask corresponding with the probability map provided
or the name of a mask that has been locally loaded.}
}
\value{
A tibble with threshold, dsc, and volume columns
}
\description{
Calculates Sørensen's–Dice coefficient and automatic volume
 estimation for a grid of thresholds.
}
\examples{
\dontrun{
gen_thresh_data(thresholds = seq(from = 0, to = 1, by = 0.01),
pmap = 'probability_map_subject_1.nii.gz',
gold_standard = 'manual_segmentation_subject_1.nii.gz',
mask = 'brain_mask_subject_1.nii.gz',
k = 8)
}
}
