% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_D.R
\name{geo_D}
\alias{geo_D}
\title{Geodesic distance between trees}
\usage{
geo_D(ths, treeH, treeS, session = "sequential", cl = 1)
}
\arguments{
\item{ths}{Trimmed matrix.}

\item{treeH}{Host phyolgeny. An object of class \code{"phylo"}.}

\item{treeS}{Symbiont phylogeny. An object of class \code{"phylo"}.}

\item{session}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"multisession"} and \code{"multicore"} (not
supported on Windows) resolves \R expressions in parallel in separate
\R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with \code{\link[future:availableCores]{future::availableCores()}}.
Note that \code{cl<=\link[=availableCores]{availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy.}
}
\value{
Geodesic distance
}
\description{
For any trimmed matrix produced with \code{\link[=trimHS_maxC]{trimHS_maxC()}},
it prunes the host-symbiont phylogenies to conform with the trimmed matrix
and computes the geodesic distance between the pruned trees.
\code{NOTE}: This function can only be used with strictly bifurcating trees.
}
\examples{
#geo_D(ths, treeH, treeS)

}
