% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_plot.R
\name{gini_plot}
\alias{gini_plot}
\title{Plot the confidence intervals of Gini coefficient}
\usage{
gini_plot(geoD, lkgd, paco, lkpaco, parafit, lkpf, ...)
}
\arguments{
\item{geoD}{Matrix produced with \code{\link[=prob_statistic]{prob_statistic()}}
for Geodesic distance.}

\item{lkgd}{Vector of statistics produced with
\code{\link[=link_freq]{link_freq()}} for Geodesic distance.}

\item{paco}{Matrix produced with \code{\link[=prob_statistic]{prob_statistic()}}
for PACo.}

\item{lkpaco}{Vector of statistics produced with
\code{\link[=link_freq]{link_freq()}} for PACo.}

\item{parafit}{Matrix produced with \code{\link[=prob_statistic]{prob_statistic()}}
for ParaFit.}

\item{lkpf}{Vector of statistics produced with
\code{\link[=link_freq]{link_freq()}} for ParaFit.}

\item{...}{}
}
\value{

}
\description{
Computes and plot the normalized Gini coefficient (G*) (Raffinetti et al.
2015) and its confidence intervals of the residual frequency distributions
of the  desired statistics: \code{"geo_D"}, \code{"paco_ss"} or
\code{"paraF"} (either all or only one of them).
}
