% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs1}
\alias{gs1}
\title{Gold Standard 1}
\usage{
gs1
}
\description{
Gold standard segmentation for subject 1.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
