% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs10}
\alias{gs10}
\title{Gold Standard 10}
\usage{
gs10
}
\description{
Gold standard segmentation for subject 10.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
