% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs11}
\alias{gs11}
\title{Gold Standard 11}
\usage{
gs11
}
\description{
Gold standard segmentation for subject 11.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
