% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs12}
\alias{gs12}
\title{Gold Standard 12}
\usage{
gs12
}
\description{
Gold standard segmentation for subject 12.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
