% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs13}
\alias{gs13}
\title{Gold Standard 13}
\usage{
gs13
}
\description{
Gold standard segmentation for subject 13.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
