% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs15}
\alias{gs15}
\title{Gold Standard 15}
\usage{
gs15
}
\description{
Gold standard segmentation for subject 15.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
