% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs2}
\alias{gs2}
\title{Gold Standard 2}
\usage{
gs2
}
\description{
Gold standard segmentation for subject 2.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
