% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs3}
\alias{gs3}
\title{Gold Standard 3}
\usage{
gs3
}
\description{
Gold standard segmentation for subject 3.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
