% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs4}
\alias{gs4}
\title{Gold Standard 4}
\usage{
gs4
}
\description{
Gold standard segmentation for subject 4.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
