% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs5}
\alias{gs5}
\title{Gold Standard 5}
\usage{
gs5
}
\description{
Gold standard segmentation for subject 5.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
