% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs6}
\alias{gs6}
\title{Gold Standard 6}
\usage{
gs6
}
\description{
Gold standard segmentation for subject 6.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
