% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs8}
\alias{gs8}
\title{Gold Standard 8}
\usage{
gs8
}
\description{
Gold standard segmentation for subject 8.
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
