% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_freq.R
\name{link_freq}
\alias{link_freq}
\title{Frequency of host-symbiont association}
\usage{
link_freq(
  x,
  fx,
  HS,
  percentile = 0.01,
  sep = "-",
  below.p = TRUE,
  res.fq = TRUE
)
}
\arguments{
\item{x}{List of trimmed matrices produced by
\code{\link[=trimHS_maxC]{trimHS_maxC()}} or
\code{\link[=trimHS_maxI]{trimHS_maxI()}}.}

\item{fx}{Vector of statistics produced with \code{\link[=geo_D]{geo_D()}},
\code{\link[=paco_ss]{paco_ss()}} or \code{\link[=paraF]{paraF()}}}

\item{HS}{Host-symbiont association matrix.}

\item{percentile}{Percentile to evaluate. Default is \code{0.01}.}

\item{sep}{Character that separates host and symbiont labels.}

\item{below.p}{Determines whether frequencies are to be computed below or
above the percentile set. Default is \code{TRUE}.}

\item{res.fq}{Determines whether a correction to avoid one-to-one
associations being overrepresented in the percentile evaluated.
If \code{TRUE} (default) a residual frequency value (observed -
expected frequency) is computed for each host-symbiont association.}
}
\value{
A dataframe with host-symbiont associations in rows. The first and
        second columns display the names of the host and symbiont terminals,
        respectively. The third column designates the host-symbiont
        association by pasting the names of the terminals, and the fourth
        column displays the frequency of occurrence of each host-symbiont
        association. If \code{res.fq = TRUE}, column 5 displays the
        corrected frequencies as a residual.
}
\description{
After applying \code{\link[=geo_D]{geo_D()}},
\code{\link[=paco_ss]{paco_ss()}} or \code{\link[=paraF]{paraF()}} to each
matrix produced by \code{\link[=trimHS_maxC]{trimHS_maxC()}}, this function
determines the frequency of each host-symbiont association occurring in a
given percentile of cases that maximize phylogenetic congruence
(or incongruence in case that \code{\link[=trimHS_maxI]{trimHS_maxI()}}
has been used).
}
\section{NOTE}{

       The \code{res.fq = TRUE} correction is recommended in tanglegrams with
       large portion of multiple (as opposed to one-to-one) host-symbiont
       associations. For future usage, frequencies of host-symbiont
       associations above a given percentile values can also be computed
       setting \code{below.p = FALSE}.
}

\examples{
# birds_mites dataset
data(birds_mites)
N = 1e+2
n = 50
TBM <- trimHS_maxC(N, bm_matrix, n, strat = "parallel", cl = 4)
GD <- geo_D(TBM, treeH = birds, treeS = mites, strat = "parallel", cl = 8)
# link frequencies with the geodesic distance
LFGD <- link_freq(TBM, GD, bm_matrix, below.p = TRUE, res.fq = FALSE)


}
