% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scattergrams.R
\name{make_scattergrams}
\alias{make_scattergrams}
\title{Make a Scatterplot of Subject-Level Thresholds with Marginal Histogram}
\usage{
make_scattergrams(tapas_model)
}
\arguments{
\item{tapas_model}{The object returned from `tapas_train`.}
}
\value{
An object of class `ggExtraPlot`. This object can be printed to show
  the plots or saved using any of the typical image-saving functions (for
  example, using `png()`` or `pdf()``).
}
\description{
Code to help visualize the subject-specific thresholds selected
  using both a scatter plot of the subject-specific best threshold from
  training data on the subject ID with a marginal histogram of the
  subject-specific thresholds. This function will help users re-fine the
  TAPAS threshold grid applied in `tapas_data`.
}
\examples{
#TODO
}
