% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_cong.R
\name{max_cong}
\alias{max_cong}
\title{Algortihm for maximizing congruence between two phylogenies}
\usage{
max_cong(
  HS,
  treeH,
  treeS,
  n,
  N,
  method = "paco",
  symmetric = FALSE,
  ei.correct = "none",
  percentile = 0.01,
  res.fq = TRUE,
  strat = "sequential",
  cl = 1
)
}
\arguments{
\item{HS}{Host-Symbiont association matrix.}

\item{treeH}{Host phyolgeny. An object of class "phylo".}

\item{treeS}{Symbiont phylogeny. An object of class "phylo".}

\item{n}{Number of unique associations.}

\item{N}{Number of runs.}

\item{method}{Specifies the desired global-fit method (GD, PACo or ParaFit).
The default is \code{PACo}.}

\item{symmetric}{Specifies the type of Procrustes superimposition. Default
is \code{FALSE}, indicates that the superposition is applied
asymmetrically (S depends on H). If \code{TRUE}, PACo is applied
symmetrically (dependency between S and H is reciprocal).}

\item{ei.correct}{Specifies how to correct potential negative eigenvalues
from the conversion of phylogenetic distances into Principal
Coordinates: \code{"none"} (the default) indicates that no correction
is required, particularly if H and S are ultrametric; \code{"sqrt.D"}
takes the element-wise square-root of the phylogenetic distances;
\code{"lingoes"} and \code{"cailliez"} apply the classical Lingoes
and Cailliez corrections, respectively.}

\item{percentile}{Percentile to evaluate (\emph{p}). Default is
\code{0.01} (1%).}

\item{res.fq}{Determines whether a correction to avoid one-to-one
associations being overrepresented in the percentile evaluated.
If \code{TRUE} (default) a residual frequency value (observed -
expected frequency) is computed for each host-symbiont association.}

\item{strat}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"parallel"} resolves \R expressions in parallel in
separate \R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with
\code{\link[parallelly:availableCores]{parallelly::availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy.}
}
\value{
A dataframe with host-symbiont associations in rows. The first and
        second columns display the names of the host and symbiont terminals,
        respectively. The third column designates the host-symbiont
        association by pasting the names of the terminals, and the fourth
        column displays the frequency of occurrence of each host-symbiont
        association in \emph{p}. If \code{res.fq = TRUE},column 5 displays
        the corrected frequencies as a residual.
}
\description{
Prunes the host (H) and symbiont (S) phylogenies to conform with trimmed
matrices and computes the given global fit method, Geodesic distances (GD),
Procrustes Approach to Cophylogeny (PACo) or ParaFit (Legendre et al. 2002)
between the pruned trees. Then, determines the frequency of each
host-symbiont association occurring in a given percentile of cases that
maximize phylogenetic congruence.
}
\section{NOTE}{

      If the \code{node.label} object in both trees contains NAs or neither
      value, an error will appear. You should make sure that all nodes
      have a value, or else remove it within the \code{"phylo"} class tree
      with \code{tree$node.label <- NULL}. For more details, see
      \code{\link[distory:dist.multiPhylo]{distory::dist.multiPhylo()}}
}

\examples{
data(nuc_pc)
N = 1e+2
n = 15
NPc <- max_cong(np_matrix, NUCtr, CPtr, n, N, method = "paco",
                symmetric = FALSE, ei.correct = "sqrt.D",
                percentile = 0.01, res.fq = FALSE,
                strat = "parallel", cl = 10)


}
