% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_f.R
\name{one2one_f}
\alias{one2one_f}
\title{Maximum n for pick unique one-to-one association over a number of runs}
\usage{
one2one_f(
  HS,
  reps = 10000,
  interval = NULL,
  strat = "sequential",
  cl = 1,
  plot = TRUE
)
}
\arguments{
\item{HS}{Host-symbiont association matrix.}

\item{reps}{Number of runs to evaluate.}

\item{interval}{Vector with the minimum and maximum \code{n} that the user
wants to test. Default is \code{"NULL"}, where a minimum \code{n}
(10% of the total associations) and a maximum \code{n} (20% of the
total associations) are automatically assigned.}

\item{strat}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"parallel"} resolves \R expressions in parallel in
separate \R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with
\code{\link[parallelly:availableCores]{parallelly::availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy.}

\item{plot}{Default is \code{"TRUE"}, plots the number of unique host-
symbiont associations for the number of runs.}
}
\value{
A data frame with the maximum number of unique associations
        (\code{n}) and the range of possible \code{n} to choose.
}
\description{
For a matrix of host-symbiont associations, it finds the maximum \code{n} for
which one-to-one unique associations can be picked in
\code{\link[=trimHS_maxC]{trimHS_maxC()}} or
\code{\link[=trimHS_maxI]{trimHS_maxI()}} over a number of runs.
}
\section{NOTE}{

         It can be used to decide the best \code{n} prior to application of
         \code{\link[=trimHS_maxC]{trimHS_maxC()}} or
         \code{\link[=trimHS_maxI]{trimHS_maxI()}}.
}

\examples{
N = 1e+2

# With plant_fungi data
data(plant_fungi)
n <- one2one_f(pf_matrix, reps = N, interval = c(15, 25), plot = TRUE)
# you can choose any n inside the range
n <- 15

}
