% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paco_ss.R
\name{paco_ss}
\alias{paco_ss}
\title{Procrustes Approach to Cophylogeny (PACo) of the host and symbiont
configurations}
\usage{
paco_ss(
  ths,
  treeH,
  treeS,
  symmetric = FALSE,
  proc.warns = FALSE,
  ei.correct = "none",
  strat = "sequential",
  cl = 1
)
}
\arguments{
\item{ths}{Trimmed matrix.}

\item{treeH}{Host phylogeny. An object of class \code{"phylo"}.}

\item{treeS}{Symbiont phylogeny. An object of class \code{"phylo"}.}

\item{symmetric}{Specifies the type of Procrustes superimposition. Default
is \code{FALSE}, indicates that the superposition is applied
asymmetrically (S depends on H). If \code{TRUE}, PACo is applied
symmetrically (dependency between S and H is reciprocal).}

\item{proc.warns}{Switches on/off trivial warnings returned when treeH and
treeS differ in size (number of tips). Default is \code{FALSE}.}

\item{ei.correct}{Specifies how to correct potential negative eigenvalues
from the conversion of phylogenetic distances into Principal
Coordinates: \code{"none"} (the default) indicates that no correction
is required, particularly if H and S are ultrametric; \code{"sqrt.D"}
takes the element-wise square-root of the phylogenetic distances;
\code{"lingoes"} and \code{"cailliez"} apply the classical Lingoes
and Cailliez corrections, respectively.}

\item{strat}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"parallel"} resolves \R expressions in parallel in
separate \R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with
\code{\link[parallelly:availableCores]{parallelly::availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy.}
}
\value{
A sum of squared residuals.
}
\description{
For any trimmed matrix produced with
\code{\link[=trimHS_maxC]{trimHS_maxC()}} or
\code{\link[=trimHS_maxI]{trimHS_maxI()}}, it prunes the host (H) & symbiont
(S) phylogenies to conform with the trimmed matrix and runs Procrustes
Approach to Cophylogeny (PACo) to produce the squared sum of residuals of
the Procrustes superimposition of the host and symbiont configurations
in Euclidean space.
}
\examples{
# birds_mites dataset
data(birds_mites)
N = 1e+2
n = 50
TBM <- trimHS_maxC(N, bm_matrix, n, strat = "parallel", cl = 4)
PACO <- paco_ss(TBM, birds, mites, symmetric = TRUE,
                  proc.warns = FALSE, ei.correct = "sqrt.D",
                  strat = "parallel", cl = 8)


}
