% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_matrix.R
\docType{data}
\name{pf_matrix}
\alias{pf_matrix}
\title{Binary association matrix between fungi species and plants}
\format{
A binary matrix with 58 rows (plant families) and 102 variables
(Sclerotiniaceae species).
}
\usage{
data(plant_fungi)
}
\description{
Presence or absence of association between 102 Sclerotiniaceae species and
58 plant families constructed with \code{\link[=assoc_mat]{assoc_mat()}}.
}
\references{
Klimov, P. B., Mironov, S. V., & OConnor, B. M. (2017).
Detecting ancient codispersals and host shifts by double dating of host and
parasite phylogenies: Application in proctophyllodid feather mites associated
with passerine birds. Evolution, 71(10), 2381-2397. \doi{10.1111/evo.13309}
}
\keyword{datasets}
