% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.R
\docType{data}
\name{plant}
\alias{plant}
\title{Plant families phylogeny}
\format{
An object of class \code{"phylo"} containing a list with the
details of the phylogenetic tree (i.e. edges, edges length, nodes,
and tips names).
}
\source{
Navaud, Olivier et al. (2018), Data from: Shifts in diversification
rates and host jump frequencies shaped the diversity of host range among
Sclerotiniaceae fungal plant pathogens, Dryad, Dataset
\doi{10.5061/dryad.7cs3g}.
}
\usage{
data(plant_fungi)
}
\description{
Phylogeny of the 58 plant families modified and used by Navaud et al. (2018).
}
\references{
Navaud, O., Barbacci, A., Taylor, A., Clarkson, J. P.,
Raffaele, S. (2018). Shifts in diversification rates and host jump
frequencies shaped the diversity of host range among Sclerotiniaceae
fungal plant pathogens. Molecular ecology, 27(5), 1309-1323.
\doi{10.1111/mec.14523}
}
\keyword{datasets}
