% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_fungi.R
\name{plant_fungi}
\alias{plant_fungi}
\title{plant_fungi dataset}
\format{
This data set compresses four objects:

\describe{
   \item{\code{pf_matrix}}{Association matrix between 102 Sclerotiniaceae
   species and 58 plant families constructed with
   \code{\link[=assoc_mat]{assoc_mat()}}.
   A binary matrix with 58 rows (plant families) and 102 variables
   (Sclerotiniaceae species).}

   \item{\code{pf_matrix}}{Information on the associations of the 102
   Sclerotiniaceae species and 58 plant families.
   A two-columns table representing associations between column 1
   (plant families) and column 2 (Sclerotiniaceae species).}

   \item{\code{pf_matrix}}{Phylogeny of the 58 plant families modified and
   used by Navaud et al. (2018).
   An object of class \code{"phylo"} containing a list with the
   details of the phylogenetic tree (i.e. edges, edges length, nodes,
   and tips names).}

   \item{\code{pf_matrix}}{Phylogeny of the 102 Sclerotiniaceae species
   constructed by
   Navaud et al. (2018).
   An object of class \code{"phylo"} containing a list with the
   details of the phylogenetic tree (i.e. edges, edges length, nodes, a
   nd tips names).}
 }
}
\source{
Navaud, Olivier et al. (2018), Data from: Shifts in diversification
rates and host jump frequencies shaped the diversity of host range among
Sclerotiniaceae fungal plant pathogens, Dryad, Dataset
\doi{10.5061/dryad.7cs3g}.
}
\usage{
data(plant_fungi)
}
\description{
Data set of 102 Sclerotiniaceae species and 58 plant families.
}
\references{
Klimov, P. B., Mironov, S. V., & OConnor, B. M. (2017).
Detecting ancient codispersals and host shifts by double dating of host and
parasite phylogenies: Application in proctophyllodid feather mites associated
with passerine birds. Evolution, 71(10), 2381-2397. \doi{10.1111/evo.13309}

Navaud, O., Barbacci, A., Taylor, A., Clarkson, J. P.,
Raffaele, S. (2018). Shifts in diversification rates and host jump
frequencies shaped the diversity of host range among Sclerotiniaceae
fungal plant pathogens. Molecular ecology, 27(5), 1309-1323.
\doi{10.1111/mec.14523}
}
\keyword{datasets}
