% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap1}
\alias{pmap1}
\title{Probability Map 1}
\usage{
pmap1
}
\description{
Probability map simulated for subject 1. Simulated from uniform(0.2, 0.4).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
