% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap10}
\alias{pmap10}
\title{Probability Map 10}
\usage{
pmap10
}
\description{
Probability map simulated for subject 10. Simulated from uniform(0.7, 0.9).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
