% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap11}
\alias{pmap11}
\title{Probability Map 11}
\usage{
pmap11
}
\description{
Probability map simulated for subject 11. Simulated from uniform(0.2, 0.8).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
