% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap12}
\alias{pmap12}
\title{Probability Map 12}
\usage{
pmap12
}
\description{
Probability map simulated for subject 12. Simulated from uniform(0.05, 0.3).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
