% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap13}
\alias{pmap13}
\title{Probability Map 13}
\usage{
pmap13
}
\description{
Probability map simulated for subject 13. Simulated from uniform(0.23, 0.3).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
