% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap14}
\alias{pmap14}
\title{Probability Map 14}
\usage{
pmap14
}
\description{
Probability map simulated for subject 14. Simulated from uniform(0.5, 0.65).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
