% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap15}
\alias{pmap15}
\title{Probability Map 15}
\usage{
pmap15
}
\description{
Probability map simulated for subject 15. Simulated from uniform(0.2, 0.6).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
