% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap2}
\alias{pmap2}
\title{Probability Map 2}
\usage{
pmap2
}
\description{
Probability map simulated for subject 2. Simulated from uniform(0.3, 0.6).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
