% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap3}
\alias{pmap3}
\title{Probability Map 3}
\usage{
pmap3
}
\description{
Probability map simulated for subject 3. Simulated from uniform(0.1, 0.8).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
