% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap4}
\alias{pmap4}
\title{Probability Map 4}
\usage{
pmap4
}
\description{
Probability map simulated for subject 4. Simulated from uniform(0.6, 0.9).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
