% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap5}
\alias{pmap5}
\title{Probability Map 5}
\usage{
pmap5
}
\description{
Probability map simulated for subject 5. Simulated from uniform(0.3, 0.7).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
