% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap6}
\alias{pmap6}
\title{Probability Map 6}
\usage{
pmap6
}
\description{
Probability map simulated for subject 6. Simulated from uniform(0.25, 0.3).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
