% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap7}
\alias{pmap7}
\title{Probability Map 7}
\usage{
pmap7
}
\description{
Probability map simulated for subject 7. Simulated from uniform(0.1, 0.3).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
