% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap8}
\alias{pmap8}
\title{Probability Map 8}
\usage{
pmap8
}
\description{
Probability map simulated for subject 8. Simulated from uniform(0.55, 0.75).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
