% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmap9}
\alias{pmap9}
\title{Probability Map 9}
\usage{
pmap9
}
\description{
Probability map simulated for subject 9. Simulated from uniform(0.15, 0.45).
This data was created by Alessandra Valcarcel and is artificial for example purposes.
}
