% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tapas.R
\name{predict_tapas}
\alias{predict_tapas}
\title{TAPAS Prediction}
\usage{
predict_tapas(pmap, model, clamp = TRUE, k = 8, verbose = TRUE)
}
\arguments{
\item{pmap}{A file path to a \code{nifti} probability map generated from an automatic segmentation
approach or the name of a probability map locally available}

\item{model}{an object of class \code{gam} used to make the
TAPAS predictions}

\item{clamp}{is \code{TRUE} by default. Uses the clamped (10th or 90th percentile)
subject-specific threshold prediction rather than the true model predicted threshold to avoid extrapolation
when the naive volume estimate falls in the tails of the model. If \code{FALSE} then even if a subject falls
in the tail the true GAM predicted threshold will be returned.}

\item{k}{Minimum number of voxels for a cluster/component. Passed to \code{\link[extrantsr]{label_mask}}.}

\item{verbose}{is \code{TRUE} by default. \code{TRUE} returns messages throughout the generating data
function. \code{FALSE} will silence comment returns.}
}
\value{
A list containing the predicted subject-specific threshold \code{subject_threshold}, the lesion segmentation mask obtained
using the predicted subject-specific threshold \code{tapas_binary_mask}, the lesion segmentation mask obtained using the
group threshold \code{group_binary_mask}.
}
\description{
This function takes a probability map for a single subject and predicts the subject
specific threshold to apply. The function will return a list of object including the TAPAS
predicted subject-specific threshold, the lesion mask produced from using this threshold, and the lesion mask
produced from using the group threshold.
}
\examples{
\dontrun{
# Put the code from other examples in here
predict_tapas(pmap, model, clamp = TRUE, k = 8, verbose = TRUE)

}
}
