% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangle_gram.R
\name{tangle_gram}
\alias{tangle_gram}
\title{Tanglegram of the host-symbiont frequencies}
\usage{
tangle_gram(
  treeH,
  treeS,
  HS,
  colscale = "diverging",
  colgrad,
  nbreaks = 50,
  fqtab,
  res.fq = TRUE,
  node.tag = TRUE,
  cexpt = 1,
  link.lwd = 1,
  link.lty = 1,
  fsize = 0.5,
  pts = FALSE,
  link.type = "curved",
  ftype = "i",
  ...
)
}
\arguments{
\item{treeH}{Host phylogeny. An object of class \code{"phylo"}.}

\item{treeS}{Symbiont phylogeny. An object of class \code{"phylo"}.}

\item{HS}{Host-symbiont association matrix.}

\item{colscale}{Choose between \code{"diverging"}, color reflects distance
from 0 (centered at 0, recommended if \code{"res.fq = TRUE"})
or \code{"sequential"}, color reflects distance from minimum value
(spanning from the min to max frequencies observed).}

\item{colgrad}{If \code{colscale =} vector defining the color gradient of the
heatmap.}

\item{nbreaks}{Number of discrete values along \code{"colgrad"}.}

\item{fqtab}{Dataframe produced with \code{\link[=link_freq()]{link_freq()}}.}

\item{res.fq}{If \code{TRUE} it processes corrected (residual) frequencies of
fqtab (column 5). If \code{FALSE} uncorrected (absolute) frequencies
(column 4) are be processed.}

\item{node.tag}{Specifies whether maximum likelihood estimators of ancestral
states are to be computed. Default is \code{TRUE}.}

\item{cexpt}{Size of color points at terminals and nodes.}

\item{link.lwd}{The line width for plotting, default to 1.}

\item{link.lty}{The line type. Line types can either be specified as an
integer (0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash,
5=longdash, 6=twodash) or as one of the character strings
\code{"blank"}, \code{"solid"}, \code{"dashed"}, \code{"dotted"},
\code{"dotdash"}, \code{"longdash"}, or \code{"twodash"}, where
\code{"blank"} uses invisible lines' (i.e., does not draw them).}

\item{fsize}{Relative font size for tip labels.}

\item{pts}{Logical value indicating whether or not to plot filled circles at
each vertex of the tree, as well as at transition points between
mapped states. Default is \code{FALSE}.}

\item{link.type}{If curved linking lines are desired, set to \code{"curved"},
if straighten linking lines are desired, do not specify the argument.}

\item{ftype}{Font type. Options are \code{"reg"}, \code{"i"}
(italics), \code{"b"} (bold) or \code{"bi"} (bold-italics).}

\item{...}{Any graphical option admissible in \code{\link[phytools:plot.cophylo]{phytools::plot.cophylo()}}}
}
\value{
A tanglegram with quantitative information displayed as heatmap.
}
\description{
Wrapper of \code{\link[phytools:plot.cophylo]{phytools::plot.cophylo()}}
is used for mapping as heatmap the host-symbiont frequencies estimated by
Random TaPas on a tanglegram. It also plots the average frequency
(or residual frequency)of occurrence of each terminal and optionally, the
fast maximum likelihood estimators of ancestral states of each node.
}
\examples{
#tangle_gram ()


}
