% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapas_predict.R
\name{tapas_predict}
\alias{tapas_predict}
\title{TAPAS Prediction}
\usage{
tapas_predict(pmap, model, clamp = TRUE, k = 8, verbose = TRUE)
}
\arguments{
\item{pmap}{A \code{character} file path to a probability map image or an object of
class \code{nifti}.}

\item{model}{The TAPAS model fit from \code{\link{tapas_train}} of class \code{gam}. This model will be
used to make subject-specific threshold predictions.}

\item{clamp}{A \code{logical} object set to \code{TRUE} by default. This setting uses the clamped
subject-specific threshold prediction rather than the prediction fit by the
TAPAS \code{model}. This only applies to volumes exceeding those at the 10th and 90th percentile
calculated using the training data. Using the clamp data avoids extrapolation when the naive volume estimate
falls in the tails of the TAPAS model. If \code{FALSE} then the the TAPAS model predicted threshold
will be used for segmentation rather than the clamped threshold.}

\item{k}{The minimum number of voxels for a cluster/component. Passed to \code{\link[extrantsr]{label_mask}}.
Segmentation clusters of size less than k are removed from the mask, volume estimation, and the
Sørensen's–Dice coefficient (DSC) calculation.}

\item{verbose}{A \code{logical} argument to print messages. Set to \code{TRUE} by default.}
}
\value{
A \code{list} containing the TAPAS predicted subject-specific threshold (\code{subject_threshold}), the lesion
segmentation mask obtained using the TAPAS predicted subject-specific threshold (\code{tapas_binary_mask}), and the
lesion segmentation mask obtained using the group threshold (\code{group_binary_mask}).
}
\description{
This function takes a probability map for a single subject and predicts the subject-specific
threshold to apply based on the TAPAS model generated from \code{\link{tapas_train}}.
The function will return a \code{list} of objects including the TAPAS predicted subject-specific threshold,
the lesion mask produced from applying this threshold, and the lesion mask
produced from using the group threshold.
}
\examples{
\dontrun{
# Put the code from other examples in here
tapas_predict(pmap, model, clamp = TRUE, k = 8, verbose = TRUE)

}
}
