% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapas_predict_par.R
\name{tapas_predict_par}
\alias{tapas_predict_par}
\title{TAPAS Prediction in Parallel}
\usage{
tapas_predict_par(cores = 1, pmap, subject_id, model, clamp = TRUE,
  k = 8, ret = FALSE, outfile = NULL, verbose = TRUE)
}
\arguments{
\item{cores}{The number of cores to use. This argument controls at most how many child processes will
be run simultaneously. The default is set to 1.}

\item{pmap}{A \code{vector} of \code{character} file paths to probability map images or a
\code{list} object with elements of class \code{nifti}.}

\item{subject_id}{A \code{vector} of subject IDs of class \code{character}. By default this is set to \code{NULL} but users must
provide an ID.}

\item{model}{The TAPAS model fit from \code{\link{tapas_train}} of class \code{gam}. This model will be
used to make subject-specific threshold predictions.}

\item{clamp}{A \code{logical} object that is \code{TRUE} by default. This setting uses the clamped
subject-specific threshold prediction rather than the prediction fit by the
TAPAS \code{model}. This only applied to volumes exceeding those at the 10th and 90th percentile
calculated using the training data. Using the clamp data avoids extrapolation when the naive volume estimate
falls in the tails of the TAPAS model. If \code{FALSE} then the the TAPAS \code{model} predicted threshold
will be used for segmentation rather than the clamped threshold. The clamping method was
used in published work.}

\item{k}{The minimum number of voxels for a cluster/component. Passed to \code{\link[extrantsr]{label_mask}}.
Segmentation clusters of size less than k are removed from the mask, volume estimation, the and
Sørensen's–Dice coefficient (DSC) calculation.}

\item{ret}{A \code{logical} argument set to \code{TRUE} by default. Returns a nested \code{list} of objects from the
function to the local R environement. If \code{FALSE} then \code{outfile} must be specified so subject data is
saved.}

\item{outfile}{Is set to \code{NULL} by default which only returns the subject-level \code{tibble} as a list
in the local R environment. To save each subject-level \code{tibble} as an R object
specify a \code{list} or \code{vector} of file paths to save with either .rds or .RData exentions included.}

\item{verbose}{A \code{logical} argument to print messages. Set to \code{TRUE} by default.}
}
\value{
A nested \code{list}. Each element in the list contains data from a subject. The subject data
is a \code{list} object containing the objects returned from \code{\link{tapas_predict}}.
}
\description{
This function wraps \code{\link{tapas_predict}} to run in parallel. This function takes
probability maps across subjects and predicts the subject specific threshold to apply based on the
TAPAS model generated from \code{\link{tapas_train}}. The function will return or save a list of objects
for each subject including the TAPAS predicted subject-specific threshold, the lesion mask produced from
applying this threshold, as well as the lesion mask produced from using the group threshold.
}
\examples{
\dontrun{
tapas_data_par(cores = 1,
thresholds = seq(from = 0, to = 1, by = 0.01),
pmap, gold_standard,
mask,
k = 8,
subject_id = NULL,
verbose = TRUE)
}
}
