% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_freq.R
\name{link_freq}
\alias{link_freq}
\title{Frequency of host-symbiont association}
\usage{
link_freq(
  x,
  fx,
  HS,
  percentile = 0.01,
  sep = "-",
  below.p = TRUE,
  res.fq = TRUE
)
}
\arguments{
\item{x}{List of trimmed matrices produced by \code{\link[=trimHS_maxC]{trimHS_maxC()}}.}

\item{fx}{Vector of statistics produced with either
\code{\link[=geo_D]{geo_D()}} or \code{\link[=paco_ss]{paco_ss()}}.}

\item{HS}{Host-symbiont association matrix.}

\item{percentile}{Percentile to evaluate. Default is \code{0.01}. The percentile
applied can be specified with \code{percentile.res.fq}, determines whether
to apply a correction to the estimated frequencies by setting a null
model in wich the occurrence of each host-symbiont association is
evenly distributed along the whole frequency distribution.}

\item{sep}{Character that separates host and symbiont labels.}

\item{below.p}{Determines whether frequencies are to be computed below or
above the percentile set. Default is \code{TRUE}.}

\item{res.fq}{Determines whether a correction to avoid one-to-one associations
being overrepresented in the percentile evaluated. If \code{TRUE} (default)
a residual frequency value (observed - expected frequency) is computed
for each host-symbiont association.}
}
\value{
A dataframe with host-symbiont associations in rows. The first and
        second columns display the names of the host and symbiont terminals,
        respectively. The third column designates the host-symbiont
        association by pasting the names of the terminals, and the fourth
        column displays the frequency of occurrence of each host-symbiont
        association. If \code{res.fq = TRUE}, column 5 displays the corrected
        frequencies as a residual.
}
\description{
After applying \code{\link[=geo_D]{geo_D()}} or \code{\link[=paco_ss]{paco_ss()}}
to each matrix produced by \code{\link[=trimHS_maxC]{trimHS_maxC()}}, this
function determines the frequency of each host-symbiont association occurring
in a given percentile of cases that maximize phylogenetic congruence (either
the fraction of lowest geodesic distances or lowest sum of squared residuals).
}
\section{NOTE}{

       The \code{res.fq = TRUE} correction is recommended in tanglegrams with
       large portion of multiple (as opposed to one-to-one) host-symbiont
       associations. For future usage, frequencies of host-symbiont
       associations above a given percentile values can also be computed
       setting \code{below.p = FALSE}.
}

\examples{
#link_freq

}
