% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkf_CI.R
\name{linkf_CI}
\alias{linkf_CI}
\title{Confidence intervals for the frequency of host-symbiont association}
\usage{
linkf_CI(
  freqfun = "geo_D",
  x,
  fx,
  c.level = 95,
  barplot = TRUE,
  col.bar = "lightblue",
  col.ci = "darkblue",
  ...
)
}
\arguments{
\item{freqfun}{Options are \code{"geo_D"}, \code{"paco_ss"} or
\code{"paraF"}, depending on which confidence intervals you want
to compute (apply to the result of \code{\link[=geo_D]{geo_D()}},
\code{\link[=paco_ss]{paco_ss()}} or \code{\link[=paraF]{paraF()}}).}

\item{x}{Matrix produced with \code{\link[=prob_statistic]{prob_statistic()}}
for Geodesic distance, PACo or ParaFit.}

\item{fx}{Vector of statistics produced with
\code{\link[=link_freq]{link_freq()}} for Geodesic distance, PACo or
ParaFit.}

\item{c.level}{Confidence interval level. Default is \code{95} (95 per cent).}

\item{barplot}{Default is \code{"TRUE"}, plots the distribution and confidence
intervals of the frequencies.}

\item{col.bar}{A vector of colors for the bars or bar components.
By default, \code{"lightblue"} is used.}

\item{col.ci}{A vector of colors for the confidence intervals arrows.
By default, \code{"darkblue"} is used.}

\item{...}{Any graphical option admissible in
\code{\link[=barplot]{barplot()}}}
}
\value{
A dataframe with the associations (columns 1 and 2), the observed
        value of the frequencies for these associations (column 3), the mean,
        the minimum and the maximum value of the frequencies (columns 4, 5
        and 6) obtained with the sets of posterior probability trees.
}
\description{
From the matrix obtained in \code{\link[=prob_statistic]{prob_statistic()}},
compute the confidence intervals for the frequencies of the host-symbiont
associations using sets of posterior probability trees.
}
\examples{
# ficus_aganoidae

}
