% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_incong.R
\name{max_incong}
\alias{max_incong}
\title{Algortihm for maximizing incongruence between two phylogenies}
\usage{
max_incong(
  HS,
  treeH,
  treeS,
  n,
  N,
  method = "paco",
  symmetric = FALSE,
  ei.correct = "none",
  percentile = 0.99,
  diff.fq = FALSE,
  strat = "sequential",
  cl = 1
)
}
\arguments{
\item{HS}{Host-Symbiont association matrix.}

\item{treeH}{Host phyolgeny. An object of class "phylo".}

\item{treeS}{Symbiont phylogeny. An object of class "phylo".}

\item{n}{Number of unique associations.}

\item{N}{Number of runs.}

\item{method}{cc}

\item{symmetric}{Specifies the type of Procrustes superimposition. Default
is \code{FALSE}, indicates that the superposition is applied
asymmetrically (S depends on H). If \code{TRUE}, PACo is applied
symmetrically (dependency between S and H is reciprocal).}

\item{ei.correct}{Specifies how to correct potential negative eigenvalues
from the conversion of phylogenetic distances into Principal
Coordinates: \code{"none"} (the default) indicates that no correction
is required, particularly if H and S are ultrametric; \code{"sqrt.D"}
takes the element-wise square-root of the phylogenetic distances;
\code{"lingoes"} and \code{"cailliez"} apply the classical Lingoes
and Cailliez corrections, respectively.}

\item{percentile}{Percentile to evaluate. Default is \code{0.01}.}

\item{diff.fq}{FALSE}

\item{strat}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"parallel"} resolves \R expressions in parallel in
separate \R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with
\code{\link[parallelly:availableCores]{parallelly::availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy.}
}
\value{
A dataframe with host-symbiont associations in rows. The first and
        second columns display the names of the host and symbiont terminals,
        respectively. The third column designates the host-symbiont
        association by pasting the names of the terminals, and the fourth
        column displays the frequency of occurrence of each host-symbiont
        association. If \code{res.fq = TRUE}, column 5 displays the
        corrected frequencies as a residual
}
\description{
Algortihm for maximizing incongruence between two phylogenies
}
\examples{
#diff.fq

}
