% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_f.R
\name{one2one_f}
\alias{one2one_f}
\title{Maximum n for pick unique one-to-one association over a number of runs}
\usage{
one2one_f(
  HS,
  reps = 10000,
  session = "sequential",
  cl = 1,
  plot = TRUE,
  na.percent = 0.2
)
}
\arguments{
\item{HS}{Host-symbiont association matrix.}

\item{reps}{Number of runs to evaluate.}
}
\value{
The maximum number of unique associations (\code{n}).
}
\description{
For a matrix of host-symbiont associations, it finds the maximum \code{n} for
which one-to-one unique associations can be picked in
\code{\link[=trimHS_maxC]{trimHS_maxC()}} over a number of runs.
}
\section{NOTE}{

         It can be used to decide the best \code{n} prior to application of
         \code{\link[=trimHS_maxC]{trimHS_maxC()}}.
}

\examples{
#one2one_f

}
