% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paraF.R
\name{paraF}
\alias{paraF}
\title{Test of host-symbiont coevolution}
\usage{
paraF(ths, treeH, treeS, ei.correct = "none", strat = "sequential", cl = 1)
}
\arguments{
\item{ths}{Trimmed matrix.}

\item{treeH}{Host phylogeny. An object of class \code{"phylo"}.}

\item{treeS}{Symbiont phylogeny. An object of class \code{"phylo"}.}

\item{ei.correct}{Specifies how to correct potential negative eigenvalues
from the conversion of phylogenetic distances into Principal
Coordinates: \code{"none"} (the default) indicates that no correction
is required, particularly if H and S are ultrametric; \code{"sqrt.D"}
takes the element-wise square-root of the phylogenetic distances;
\code{"lingoes"} and \code{"cailliez"} apply the classical Lingoes and
Cailliez corrections, respectively.}

\item{strat}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"parallel"} resolves \R expressions in parallel in
separate \R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with \code{\link[parallelly:availableCores]{parallelly::availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy}
}
\value{
A number object with the statistic of the global host-symbiont test
        for any trimmed matrix.
}
\description{
Test of host-symbiont coevolution
}
\examples{
#paraF()
}
