% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_statistic.R
\name{prob_statistic}
\alias{prob_statistic}
\title{Frequencies of the associations for the posterior probabilistic trees}
\usage{
prob_statistic(
  ths,
  HS,
  mTreeH,
  mTreeS,
  freqfun = "geo_D",
  fx,
  percentile = 0.01,
  res.fq = TRUE,
  below.p = TRUE,
  symmetric = FALSE,
  ei.correct = "none",
  proc.warns = FALSE,
  strat = "sequential",
  cl = 1
)
}
\arguments{
\item{ths}{List of trimmed matrices produced by
\code{\link[=trimHS_maxC]{trimHS_maxC()}}.}

\item{HS}{Host-Symbiont association matrix.}

\item{mTreeH}{Number X of posterior-probabilistic trees of host.}

\item{mTreeS}{Number X of posterior-probabilistic trees of symbiont.}

\item{freqfun}{Options are \code{"geo_D"}, \code{"paco_ss"} or
\code{"paraF"}, depending on which confidence intervals you want to
compute (apply to the result of \code{\link[=geo_D]{geo_D()}},
\code{\link[=paco_ss]{paco_ss()}} or \code{\link[=paraF]{paraF()}})}

\item{fx}{Vector of statistics produced with
\code{\link[=link_freq]{link_freq()}} for Geodesic distance, PACo or
ParaFit.}

\item{percentile}{Percentile to evaluate. Default is \code{0.01}.}

\item{res.fq}{Determines whether a correction to avoid one-to-one
associations being overrepresented in the percentile evaluated.
If \code{TRUE} (default) a residual frequency value (observed -
expected frequency) is computed for each host-symbiont association.}

\item{below.p}{Determines whether frequencies are to be computed below or
above the percentile set. Default is \code{TRUE}.}

\item{symmetric}{Specifies the type of Procrustes superimposition. Default
is \code{FALSE}, indicates that the superposition is applied
asymmetrically (S depends on H). If \code{TRUE}, PACo is applied
symmetrically (dependency between S and H is reciprocal).}

\item{ei.correct}{Specifies how to correct potential negative eigenvalues
from the conversion of phylogenetic distances into Principal
Coordinates: \code{"none"} (the default) indicates that no correction
is required, particularly if H and S are ultrametric; \code{"sqrt.D"}
takes the element-wise square-root of the phylogenetic distances;
\code{"lingoes"} and \code{"cailliez"} apply the classical Lingoes
and Cailliez corrections, respectively.}

\item{proc.warns}{Switches on/off trivial warnings returned when treeH and
treeS differ in size (number of tips). Default is \code{FALSE}.}

\item{strat}{Strategy you want to work with. Default is \code{"sequential"},
resolves \R expressions sequentially in the current \R
process. If \code{"parallel"} resolves \R expressions in parallel in
separate \R sessions running in the background.}

\item{cl}{Number of cluster the user wants to use. Check how many CPUs/cores
your computer has with
\code{\link[parallelly:availableCores]{parallelly::availableCores()}}.
Default is \code{cl = 1} for \code{"sequential"} strategy.}
}
\value{
A matrix with the value of the statistics for each of the
        probabilistic trees.
}
\description{
Computes the frequencies of the associations for each of the probabilistic
trees from the statistics of \code{\link[=geo_D]{geo_D()}},
\code{\link[=paco_ss]{paco_ss()}} or \code{\link[=paraF]{paraF()}}).
}
