% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapas_data_par.R
\name{tapas_data_par}
\alias{tapas_data_par}
\title{Generate TAPAS Data in Parallel}
\usage{
tapas_data_par(cores = 1, thresholds = seq(from = 0, to = 1, by =
  0.01), pmap, gold_standard, mask, k = 8, subject_id = NULL,
  verbose = TRUE)
}
\arguments{
\item{cores}{The number of cores to use, i.e. at most how many child processes will be run
simultaneously. Default is set to 1.}

\item{thresholds}{A \code{vector} of thresholds to use for calculation of Sørensen's–Dice coefficient (DSC)
and automatic volume. The default and grid applied in the published work is 0 to 1. Threshold
values must be between 0 and 1.}

\item{pmap}{A file path to a \code{nifti} probability map generated from an automatic segmentation
approach or the name of a list of probability map locallys available.}

\item{gold_standard}{A vector of file paths to the \code{nifti} gold standard segmentation approach corresponding
with the probability map provided or the name of a list of gold standard segmentations that has been
locally loaded. The gold standard segmentation is used to compare the thresholded probability map
image with using Sørensen's–Dice coefficient.}

\item{mask}{A vector file paths to a \code{nifti} brain mask corresponding with the probability map provided
or the name of a list of masks that has been locally loaded.}

\item{k}{Minimum number of voxels for a cluster/component. Passed to \code{\link[extrantsr]{label_mask}}.}

\item{subject_id}{is \code{NULL} by default. Must be a \code{vector} of \code{character} strings unique subject IDs for
which the threshold data is being generated.}

\item{verbose}{is \code{TRUE} by default. \code{TRUE} returns messages throughout the generating data function.
\code{FALSE} will silence comment returns.}
}
\value{
A list of the \code{tibble}s returned from \code{\link{gen_tapas_data}} for each subject.
}
\description{
This function wraps the \code{\link{gen_tapas_data}} to run in parallel and create the
training data from probability maps generated using an automatic segmentation method in order generate
data necessary to train the TAPAS model. For a grid of threhsolds provided it calculates Sørensen's–Dice coefficient and automatic volume
estimation for a grid of thresholds.
}
\examples{
\dontrun{
tapas_data_par(cores = 1,
thresholds = seq(from = 0, to = 1, by = 0.01),
pmap, gold_standard,
mask,
k = 8,
subject_id = NULL,
verbose = TRUE)
}
}
