% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapas_train.R
\name{tapas_train}
\alias{tapas_train}
\title{TAPAS Model Training}
\usage{
tapas_train(data, dsc_cutoff = 0.03, verbose = TRUE)
}
\arguments{
\item{data}{Data resulting from \code{\link{tapas_data}}. The \code{data} should be a \code{tibble} or
\code{data.frame} containing binded subject data or a \code{list} object with subject data in each element.
Data from these subjects will be used for model training.}

\item{dsc_cutoff}{The Sørensen's–Dice coefficient (DSC) value to use as a cutoff for training inclusion.
By default 0.03 is used. This must be a single value between 0 and 1. Only training subjects with a subject-specific
threshold estimate resulting in Sørensen's–Dice coefficient (DSC) greater than or equal to the \code{dsc_cutoff}
will be included in training the TAPAS model.}

\item{verbose}{A \code{logical} argument to print messages. Set to \code{TRUE} by default.}
}
\value{
A \code{list} with the TAPAS model (\code{tapas_model}) of class \code{gam} and
a \code{tibble} with the clamp information (\code{clamp_data}). The clamp information contains the
TAPAS-predicted smallest and largest threshold to be applied by using estimates related to the volume at the
10th and 90th percentile.
}
\description{
This function trains the TAPAS model using all binded subject-level \code{tibble}s produced
from the \code{\link{tapas_data}} function. The TAPAS model is fit and clamp data is calculated. The clamp data
contains the predicted threshold when using the 10th and 90th percentile volume from training data.
}
\examples{
\dontrun{
# Put tapas_data function here first then run tapas_train. Fix later.
tapas_train(data = data, dsc_cutoff = 0.03, verbose = TRUE)
}
}
