% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_tapas.R
\name{train_tapas}
\alias{train_tapas}
\title{TAPAS Model Training}
\usage{
train_tapas(data, dsc_cutoff = 0.03, verbose = TRUE)
}
\arguments{
\item{data}{Data resulting from \code{\link{gen_tapas_data}}. May be in a binded
\code{tibble} or \code{data.frame} form from all subjects to include for
training or a \code{list} with all subject data in unique elements.}

\item{dsc_cutoff}{is 0.03 by default. Must be a single value between 0 and 1. This is the DSC value that will be used as a cutoff.
Only subjects with DSC greater than or equal to dsc_cutoff will be included in training the TAPAS model..}

\item{verbose}{is TRUE by default. TRUE returns messages throughout the generating data function.
FALSE will silence comment returns.}
}
\value{
A \code{list} with the TAPAS model (tapas_model) in the first element of the list of class \code{gam} and the
second element is a \code{tibble} with the clamp volume and threshold information (clamp_data).
}
\description{
This function trains the TAPAS model as a \code{tibble} or \code{data.frame}
produced by an element from the output of the function \code{\link{gen_tapas_data}}.
}
\examples{
\dontrun{
# Put gen_tapas_data function here first then run train_tapas. Fix later.
train_tapas(data = data, dsc_cutoff = 0.03, verbose = TRUE)
}
}
