% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimHS_maxI.R
\name{trimHS_maxI}
\alias{trimHS_maxI}
\title{Trims the H-S association matrix maximizing the incongruence}
\usage{
trimHS_maxI(N, HS, n, check.unique = TRUE)
}
\arguments{
\item{N}{Number of runs.}

\item{HS}{Host-Symbiont association matrix.}

\item{n}{Number of unique associations.}

\item{check.unique}{if \code{TRUE} discards duplicated trimmed matrices. This
alternative is recommended if \code{n} is small, because the
probability of obtaining the same trimmed matrix in different runs
increases as \code{n} decreases.}
}
\value{
A list of the N trimmed matrices.
}
\description{
For N runs, it randomly chooses \code{n} unique one-to-one associations and
trims the H-S association matrix to exclude the configurations with only one
H or one S.
}
\examples{
# With birds_mites dataset
# data(birds_mites)
N = 1e+2
n = 50
TBMI <- trimHS_maxI(N, bm_matrix, n, check.unique = TRUE)

}
